Добавил контроллер и перенёс в него рендеринг маршрута из main.js
