+++ src/main.js
@@ -2,88 +2,20 @@
 import TripInfoComponent from "./components/trip-info.js";
 import TripTabsComponent from "./components/trip-tabs.js";
 import TripFiltersComponent from "./components/trip-filters.js";
-import TripSortComponent from "./components/trip-sort.js";
-import TripDaysComponent from "./components/trip-days.js";
-import TripDaysItemComponent from "./components/trip-days-item.js";
-import TripEventsMsgComponent from "./components/trip-events-msg.js";
-import TripEventsItemEditComponent from "./components/trip-events-item-edit.js";
-import TripEventsItemComponent from "./components/trip-events-item.js";
-import {render, replace, RenderPosition} from "./utils/render.js";
+
+// Контроллеры
+import TripController from "./controllers/trip.js";
+
+// Утилиты
+import {render, RenderPosition} from "./utils/render.js";
 
 // Моки
-import {generateTripEvents, getEventsForDate} from "./mock/trip-events-item.js";
-import {getTripDates} from "./mock/trip-days-item.js";
+import {generateTripEvents} from "./mock/trip-events-item.js";
 
 // Константы
 const EVENTS_COUNT = 20;
 
 const events = generateTripEvents(EVENTS_COUNT);
-
-// Отрисовка точки маршрута и формы создания/редактирования
-const renderTripEventsItem = (tripEventsListElement, eventsItem) => {
-  const replaceEventToEdit = () => {
-    replace(tripEventsItemEditComponent, tripEventsItemComponent);
-  };
-
-  const replaceEditToEvent = () => {
-    replace(tripEventsItemComponent, tripEventsItemEditComponent);
-  };
-
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscKey) {
-      replaceEditToEvent();
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  const tripEventsItemComponent = new TripEventsItemComponent(eventsItem);
-  const tripEventsItemEditComponent = new TripEventsItemEditComponent(eventsItem);
-
-  tripEventsItemComponent.setEventRollupBtnClickHandler(() => {
-    replaceEventToEdit();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
-
-  tripEventsItemEditComponent.setSubmitHandler((evt) => {
-    evt.preventDefault();
-    replaceEditToEvent();
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  });
-
-  render(tripEventsListElement, tripEventsItemComponent, RenderPosition.BEFOREEND);
-};
-
-// Отрисовка маршрута
-const renderTripEvents = (tripEventsElement, events) => {
-  const noEvents = (events.length === 0);
-
-  if (noEvents) {
-    render(tripEventsElement, new TripEventsMsgComponent(), RenderPosition.BEFOREEND);
-    return;
-  }
-
-  render(tripEventsElement, new TripSortComponent(), RenderPosition.BEFOREEND);
-  render(tripEventsElement, new TripDaysComponent(), RenderPosition.BEFOREEND);
-
-  const tripDaysElement = tripEventsElement.querySelector(`.trip-days`);
-
-  const dates = getTripDates(events);
-
-  dates.forEach((date, index) => {
-    const tripDaysItemComponent = new TripDaysItemComponent(date, index);
-
-    render(tripDaysElement, tripDaysItemComponent, RenderPosition.BEFOREEND);
-
-    const tripEventsListElement = tripDaysItemComponent.getElement().querySelector(`.trip-events__list`);
-
-    const eventsForDate = getEventsForDate(events, date);
-
-    eventsForDate.forEach((eventsItem) =>
-      renderTripEventsItem(tripEventsListElement, eventsItem));
-  });
-};
 
 const tripMainElement = document.querySelector(`.trip-main`);
 
@@ -98,4 +30,6 @@ render(tripControlsElement, new TripFiltersComponent(), RenderPosition.BEFOREEND
 
 const tripEventsElement = document.querySelector(`.trip-events`);
 
-renderTripEvents(tripEventsElement, events);
+const tripController = new TripController(tripEventsElement);
+
+tripController.render(events);
